Remove `keep_prefix` default behavior... actually remove the feature entirely. (#3)
This feature was not good at all. Felt like it was a good idea. It was complicated and I ended up disabling it. Very bad as a default, bad generally as an inclusion. If a user wants to keep services just specify them all; this is for applicaiton docker builds anyways and really should be tested.

Also this PR adds automated tests because I hadn't touched this in a few days and already forgot how to run the destructive tests.